migrations and seeding done by separate flask app
